--- config/doctrine/ConsultationBeneficiaire.orm.yml
@@ -1,28 +1,0 @@
-App\Entity\ConsultationBeneficiaire:
-    type: entity
-    table: "consultationbeneficiaire"
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        createdAt:
-            type: datetime
-            column: createdAt
-            gedmo:
-                timestampable:
-                    on: create
-    manyToOne:
-        membre:
-            targetEntity: App\Entity\Membre
-            inversedBy: consultationsBeneficiaires
-            joinColumn:
-                nullable: false
-        beneficiaire:
-            targetEntity: App\Entity\Attributes\Beneficiaire
-            inversedBy: consultationsBeneficiaires
-            joinColumn:
-                nullable: false
-    lifecycleCallbacks: {  }
