+++ config/doctrine/Membre.orm.yml
@@ -45,7 +45,7 @@ App\Entity\Membre:
             cascade: [ "persist", "remove" ]
             mappedBy: membre
         consultationsBeneficiaires:
-            targetEntity: App\Entity\ConsultationBeneficiaire
+            targetEntity: App\Entity\Attributes\ConsultationBeneficiaire
             cascade: ["remove"]
             mappedBy: membre
         evenements:
