+++ src/Manager/ConsultationBeneficiaireManager.php
@@ -3,7 +3,7 @@
 namespace App\Manager;
 
 use App\Entity\Attributes\Beneficiaire;
-use App\Entity\ConsultationBeneficiaire;
+use App\Entity\Attributes\ConsultationBeneficiaire;
 use App\Entity\Membre;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
@@ -47,7 +47,7 @@ class ConsultationBeneficiaireManager
     {
         $result = $this->em->createQueryBuilder()
             ->select('cb')
-            ->from('App:ConsultationBeneficiaire', 'cb')
+            ->from(ConsultationBeneficiaire::class, 'cb')
             ->where('cb.beneficiaire = '.$beneficiaire->getId())
             ->andWhere('cb.membre = '.$membre->getId())
             ->getQuery()
